+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.4.2-preview.1] - 2021-02-08
+- Add Extension_CapsuleCollider2D script.
+- Add GetPositionTop and GetPositionBottom method in Extension_CapsuleCollider2D.
+
 ## [0.4.1-preview.1] - 2021-01-17
 - Add Dictionary extension TryAdd.